Ifpack2::RILUK: Refactor setParameters to remove duplicated code
@trilinos/ifpack2
